Update next to canary
- Update color of the geojson on map
- Move functions to utils
