+++ next.config.js
@@ -1,6 +1,7 @@
 const withExportImages = require('next-export-optimize-images');
 
 const nextConfig = withExportImages({
+    output: 'export',
     reactStrictMode: true,
     images: {
         // domain: ['localhost: 8020'],
