+++ package.json
@@ -4,9 +4,9 @@
     "private": true,
     "scripts": {
         "dev": "next dev -p 3020",
-        "build": "next build",
+        "build": "next build && next-export-optimize-images",
         "start": "next start",
-        "export": "next build && next export && next-export-optimize-images",
+        "export": "next build && next-export-optimize-images",
         "generate": "graphql-codegen --config codegen.yml && eslint --fix src/generated/types.ts",
         "generate:i18n": "i18next 'src/**/*.{jsx,tsx}' -o 'public/locales/$LOCALE/$NAMESPACE.json'",
         "lint": "next lint",
@@ -20,7 +20,7 @@
         "i18next": "^22.5.1",
         "leaflet": "^1.9.4",
         "leaflet-gesture-handling": "^1.2.2",
-        "next": "^13.4.5",
+        "next": "^13.4.7-canary.0",
         "next-i18next": "^13.3.0",
         "next-language-detector": "^1.0.2",
         "papaparse": "^5.4.1",
