+++ src/components/RawInput/index.tsx
@@ -46,7 +46,6 @@ function RawInput<const N>(props: Props<N>) {
                 styles.rawInput,
                 className,
             )}
-            // FIXME: do we even need to pass name?
             name={isDefined(name) ? String(name) : undefined}
             onChange={handleChange}
             value={value ?? ''}
