+++ src/pages/[locale]/get-involved/index.tsx
@@ -5,7 +5,6 @@ import { _cs } from '@togglecorp/fujs';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 
 import Hero from 'components/Hero';
-import Heading from 'components/Heading';
 import Button from 'components/Button';
 import Link from 'components/Link';
 import Section from 'components/Section';
